docs(directives): updated usage
udated usage on directives and renamed motion scrolling
