+++ packages/vuetifyjs.com/src/data/pages/motion/Scroll.json
@@ -16,9 +16,7 @@
       "children": [
         {
           "type": "api",
-          "value": [
-            "$vuetify"
-          ]
+          "value": [ "$vuetify" ]
         }
       ]
     },
