+++ packages/vuetifyjs.com/src/examples/resizing/usage.vue
