--- packages/vuetifyjs.com/src/examples/scrolling/default.vue
@@ -1,20 +1,0 @@
-<template>
-  <v-layout v-scroll="onScroll" column align-center justify-center>
-    <v-subheader>Offset Top</v-subheader>
-    {{ offsetTop }}
-  </v-layout>
-</template>
-
-<script>
-  export default {
-    data: () => ({
-      offsetTop: 0
-    }),
-
-    methods: {
-      onScroll (e) {
-        this.offsetTop = window.pageYOffset || document.documentElement.scrollTop
-      }
-    }
-  }
-</script>
