+++ packages/vuetifyjs.com/src/lang/en/directives/Resizing.json
@@ -2,8 +2,7 @@
   "header": "# Resize directive",
   "headerText": "The `v-resize` directive can be used for calling specific functions when the window resizes.",
   "examples": {
-    "default": {
-      "header": "Default",
+    "usage": {
       "desc": "Resize your window and observe the values change..",
       "uninverted": true
     }
