+++ packages/vuetifyjs.com/src/lang/en/directives/Ripples.json
@@ -2,8 +2,7 @@
   "header": "# Ripple directive",
   "headerText": "The `v-ripple` directive is used to show action from a user. It can be applied to any block level element. Numerous components come with the ripple directive built in, such as the `v-btn`, `v-tabs-item` and many more.",
   "examples": {
-    "buttons": {
-      "header": "Default",
+    "usage": {
       "desc": "Buttons by default have ripples enabled. This can be removed with the prop `:ripple=\"false\"`",
       "uninverted": true
     },
