+++ packages/vuetifyjs.com/src/lang/en/directives/Scrolling.json
@@ -2,13 +2,12 @@
   "header": "# Scrolling directive",
   "headerText": "The `v-scroll` directive allows you to provide callbacks when the window or a specifically defined element are scrolled.",
   "examples": {
-    "default": {
-      "header": "## Default",
+    "usage": {
       "desc": "The default behavior is to bind to the window. If no additional configuration options are needed, you can simply pass your callback function.",
       "uninverted": true
     },
     "options": {
-      "header": "## Scroll with options",
+      "header": "### Scroll with options",
       "desc": "For a more fine tuned approach, you can designate the target to bind the scroll event listener.",
       "uninverted": true
     }
