+++ packages/vuetifyjs.com/src/lang/en/directives/TouchSupport.json
@@ -2,8 +2,7 @@
   "header": "# Touch Support",
   "headerText": "The <code>v-touch</code> directive allows you to capture swipe gestures and apply directional callbacks.",
   "examples": {
-    "default": {
-      "header": "Default",
+    "usage": {
       "desc": "On a mobile device, try swiping in various directions.",
       "uninverted": true
     }
