+++ packages/vuetifyjs.com/src/lang/en/motion/Scroll.json
@@ -1,5 +1,5 @@
 {
-  "header": "# Scrolling",
+  "header": "# Scroll",
   "headerText": "You can programatically trigger scrolling in your application by using the `goTo` method found on the `$vuetify` object. This method supports several different types of target selectors, and options including smooth scrolling using built-in easing functions.",
   "firstHeader": "## First",
   "secondHeader": "## Second",
